+++ packages/flutter_riverpod/test/framework_test.dart
@@ -794,11 +794,7 @@ class _InitStateState extends ConsumerState<_InitState> {
 }
 
 final class _TestVsync implements Vsync {
-  _TestVsync({
-    required void Function()? Function(Task task) scheduleRefresh,
-    required void Function()? Function(Task task) scheduleDispose,
-  }) : _scheduleRefresh = scheduleRefresh,
-       _scheduleDispose = scheduleDispose;
+  _TestVsync({required this._scheduleRefresh, required this._scheduleDispose});
 
   final void Function()? Function(Task task) _scheduleRefresh;
   final void Function()? Function(Task task) _scheduleDispose;