+++ packages/riverpod/lib/src/core/async_value.dart
@@ -688,13 +688,8 @@ final class AsyncData<ValueT> extends AsyncResult<ValueT> {
     @internal DataKind? kind,
   }) : this._((value, kind: kind, source: null), loading: null, error: null);
 
-  const AsyncData._(
-    this._value, {
-    required _ErrorRecord? error,
-    required _LoadingRecord? loading,
-  }) : _loading = loading,
-       _error = error,
-       super._();
+  const AsyncData._(this._value, {required this._error, required this._loading})
+    : super._();
 
   @override
   final _LoadingRecord? _loading;
@@ -747,11 +742,9 @@ final class AsyncLoading<ValueT> extends AsyncValue<ValueT> {
 
   const AsyncLoading._(
     this._loading, {
-    required _DataRecord<ValueT>? value,
-    required _ErrorRecord? error,
-  }) : _value = value,
-       _error = error,
-       super._();
+    required this._value,
+    required this._error,
+  }) : super._();
 
   @override
   final _LoadingRecord _loading;
@@ -835,11 +828,9 @@ final class AsyncError<ValueT> extends AsyncResult<ValueT> {
 
   const AsyncError._(
     this._error, {
-    required _DataRecord<ValueT>? value,
-    required _LoadingRecord? loading,
-  }) : _value = value,
-       _loading = loading,
-       super._();
+    required this._value,
+    required this._loading,
+  }) : super._();
 
   @override
   final _LoadingRecord? _loading;