+++ packages/riverpod/lib/src/core/provider_container.dart
@@ -204,17 +204,16 @@ class ProviderDirectory implements _PointerBase {
       currentContainer: currentContainer,
       targetContainer: targetContainer,
       inherit: (target) => target._pointerManager.upsertPointer(provider),
-      scope:
-          ({override}) => $ProviderPointer(
-            targetContainer: currentContainer,
-            providerOverride:
-                override == null ||
-                        provider.from !=
-                            null //
-                    ? null
-                    : TransitiveProviderOverride(override),
-            origin: provider,
-          ),
+      scope: ({override}) => $ProviderPointer(
+        targetContainer: currentContainer,
+        providerOverride:
+            override == null ||
+                provider.from !=
+                    null //
+            ? null
+            : TransitiveProviderOverride(override),
+        origin: provider,
+      ),
     );
   }
 
@@ -357,11 +356,10 @@ class ProviderPointerManager {
       return;
     }
 
-    final familyPointer =
-        familyPointers[from] ??= ProviderDirectory.empty(
-          container._root ?? container,
-          familyOverride: null,
-        );
+    final familyPointer = familyPointers[from] ??= ProviderDirectory.empty(
+      container._root ?? container,
+      familyOverride: null,
+    );
 
     familyPointer.addProviderOverride(override, targetContainer: container);
   }
@@ -434,11 +432,7 @@ class ProviderPointerManager {
                 familyPointer.pointers.values.map((e) => e.targetContainer),
               );
             case $ProviderBaseImpl():
-              return [
-                if (readPointer(dependency)?.targetContainer
-                    case final container?)
-                  container,
-              ];
+              return [?readPointer(dependency)?.targetContainer];
           }
         });
 
@@ -470,9 +464,9 @@ class ProviderPointerManager {
       scope: ({override}) {
         final familyOverride =
             override ==
-                    null //
-                ? null
-                : TransitiveFamilyOverride(override);
+                null //
+            ? null
+            : TransitiveFamilyOverride(override);
 
         final parent = container.parent?._pointerManager.familyPointers[family];
 
@@ -942,17 +936,13 @@ final class ProviderContainer implements MutationTarget {
       }
     }
 
-    _pointerManager =
-        parent != null
-            ? ProviderPointerManager.from(parent, overrides, container: this)
-            : ProviderPointerManager(
-              overrides,
-              container: this,
-              orphanPointers: ProviderDirectory.empty(
-                this,
-                familyOverride: null,
-              ),
-            );
+    _pointerManager = parent != null
+        ? ProviderPointerManager.from(parent, overrides, container: this)
+        : ProviderPointerManager(
+            overrides,
+            container: this,
+            orphanPointers: ProviderDirectory.empty(this, familyOverride: null),
+          );
 
     // Mutate the parent & global state only at the very end.
     // This ensures that if an error is thrown, the parent & global state
@@ -1061,10 +1051,7 @@ final class ProviderContainer implements MutationTarget {
   Future<void> pump() async {
     final a = scheduler.pendingFuture;
 
-    await Future.wait<void>([
-      if (a != null) a,
-      if (parent case final parent?) parent.pump(),
-    ]);
+    await Future.wait<void>([?a, if (parent case final parent?) parent.pump()]);
   }
 
   /// Reads a provider without listening to it and returns the currently
@@ -1120,10 +1107,9 @@ final class ProviderContainer implements MutationTarget {
   /// ref.read(myFamily(0));
   /// ```
   Iterable<ProviderReference> allProviders({Family? family}) {
-    final providers =
-        family != null
-            ? _pointerManager.listFamilyProviders(family)
-            : _pointerManager.listProviders();
+    final providers = family != null
+        ? _pointerManager.listFamilyProviders(family)
+        : _pointerManager.listProviders();
 
     if (container.parent case final parent?) {
       return providers.followedBy(parent.allProviders(family: family));