+++ packages/riverpod/test/src/core/provider_listenable_transformer_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use_from_same_package
+
 import 'dart:async';
 
 import 'package:mockito/mockito.dart';