+++ packages/riverpod/test/src/matrix/async_notifier_provider.dart
@@ -140,11 +140,7 @@ abstract class TestAsyncNotifier<ValueT> implements $AsyncNotifier<ValueT> {}
 
 class DeferredAsyncNotifier<ValueT> extends AsyncNotifier<ValueT>
     implements TestAsyncNotifier<ValueT> {
-  DeferredAsyncNotifier(
-    this._create, {
-    bool Function(AsyncValue<ValueT>, AsyncValue<ValueT>)? updateShouldNotify,
-    this.arg,
-  }) : _updateShouldNotify = updateShouldNotify;
+  DeferredAsyncNotifier(this._create, {this._updateShouldNotify, this.arg});
 
   final Object? arg;