+++ packages/riverpod/test/src/matrix/notifier_provider.dart
@@ -123,11 +123,7 @@ abstract class TestNotifier<ValueT> implements $Notifier<ValueT> {
 
 class DeferredNotifier<ValueT> extends Notifier<ValueT>
     implements TestNotifier<ValueT> {
-  DeferredNotifier(
-    this._create, {
-    bool Function(ValueT, ValueT)? updateShouldNotify,
-    this.arg,
-  }) : _updateShouldNotify = updateShouldNotify;
+  DeferredNotifier(this._create, {this._updateShouldNotify, this.arg});
 
   final Object? arg;