+++ packages/riverpod_lint_flutter_test/test/lints/provider_dependencies/missing_dependencies2.provider_dependencies-126.fix.dart
@@ -30,6 +30,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:riverpod_annotation/experimental/scope.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
+
 import '../../test_annotation.dart';
 import 'another.dart' as import_alias;
 import 'another.dart';