+++ website/i18n/ru/docusaurus-plugin-content-docs/current/concepts2/auto_dispose/invalidate_family_example/codegen.dart
@@ -1,0 +1,27 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+part 'codegen.g.dart';
+
+late WidgetRef ref;
+
+/* SNIPPET START */
+@riverpod
+String label(Ref ref, String userName) {
+  return 'Hello $userName';
+}
+
+// ...
+
+void onTap() {
+  // {@template invalidateAll}
+  // Инвалидировать все возможные комбинации параметров этого provider.
+  // {@endtemplate}
+  ref.invalidate(labelProvider);
+  // {@template invalidate}
+  // Инвалидировать только конкретную комбинацию параметров.
+  // {@endtemplate}
+  ref.invalidate(labelProvider('John'));
+}
+
+/* SNIPPET END */